Conditional statements for different users

+++ torquay_sunday_league/routes.py
@@ -8,7 +8,10 @@ import datetime
 
 @app.route("/")
 def home():
-    return render_template("index.html")
+    user = session["user"]
+    if user.team_managed != "":
+        team = Team.query.filter_by(user.team_managed=user.team_managed).first()
+    return render_template("index.html", user=user)
 
 
 @app.route("/teams")
@@ -18,20 +21,25 @@ def teams():
     return render_template("teams.html", teams=teams)
 
 
-@app.route("/create_team", methods=["GET", "POST"])
-def create_team():
+@app.route("/create_team/<username>", methods=["GET", "POST"])
+def create_team(username):
+    user = User.query.filter_by(username=username).first()
     if request.method == "POST":
-        team = Team(
-            team_name=request.form.get("team_name"),
-            team_no_of_players=0,
-            team_colour=request.form.get("team_colour"),
-            team_location=request.form.get("team_location"),
-            team_created_by=session["user"]
-            )
-        db.session.add(team)
-        db.session.commit()
-        return redirect(url_for("teams"))
-    return render_template("create_team.html")
+        if session["user"] == "ryanmcnally93" or user.team_managed == "":
+            team = Team(
+                team_name=request.form.get("team_name"),
+                team_no_of_players=0,
+                team_colour=request.form.get("team_colour"),
+                team_location=request.form.get("team_location"),
+                team_created_by=session["user"]
+                )
+            db.session.add(team)
+            db.session.commit()
+            user.team_managed == team.team_name
+            return redirect(url_for("teams"))
+        else:
+            flash("A user may only register one team.")
+    return render_template("create_team.html", username=session["user"])
 
 
 @app.route("/register", methods=["GET", "POST"])
@@ -63,7 +71,7 @@ def register():
         # put the new user into 'session' cookie
         session["user"] = request.form.get("username").lower()
         flash("Registration Successful!")
-        return redirect(url_for("profile", username=session["user"]))
+        return redirect(url_for("create_team", username=session["user"]))
     return render_template("register.html")
 
 
@@ -98,11 +106,12 @@ def log_in():
 @app.route("/profile/<username>", methods=["GET", "POST"])
 def profile(username):
     user = User.query.filter_by(username=username).first()
+    team = user.team_managed
     username = session["user"]
 
     # Checking for session cookie
     if session["user"]:
-        return render_template("user_profile.html", username=username, user=user)
+        return render_template("user_profile.html", username=username, user=user, team=team)
     
     # If no session cookie, we return to log_in page
     return redirect(url_for("log_in"))
@@ -119,22 +128,28 @@ def log_out():
 def edit_team(team_id):
     team = Team.query.get_or_404(team_id)
     if request.method == "POST":
-        team.team_name = request.form.get("team_name")
-        team.team_no_of_players = len(Player.query.get(team_id))
-        team.team_colour = request.form.get("team_colour")
-        team.team_location = request.form.get("team_location")
-        db.session.commit()
-        # Change redirect for team profile page
-        return redirect(url_for("teams"))
+        if session["user"] == team.team_created_by:
+            team.team_name = request.form.get("team_name")
+            team.team_no_of_players = len(Player.query.get(team_id))
+            team.team_colour = request.form.get("team_colour")
+            team.team_location = request.form.get("team_location")
+            db.session.commit()
+            # Change redirect for team profile page
+            return redirect(url_for("teams"))
+        else:
+            flash("Only the creator of this team may edit it.")
     return render_template("edit_team.html", team=team)
 
 
 @app.route("/delete_team/<int:team_id>")
 def delete_team(team_id):
     team = Team.query.get_or_404(team_id)
-    db.session.delete(team)
-    db.session.commit()
-    return redirect(url_for("teams"))
+    if session["user"] == team.team_created_by:
+        db.session.delete(team)
+        db.session.commit()
+        return redirect(url_for("teams"))
+    else:
+        flash("Only the creator of this team may delete it.")
 
 
 @app.route("/players/<int:id>")
@@ -150,14 +165,17 @@ def edit_player(player_id, team_id):
     teams = list(Team.query.order_by(Team.team_name).all())
     team = Team.query.get_or_404(team_id)
     if request.method == "POST":
-        player.player_kit_number=request.form.get("player_kit_number")
-        player.player_name = request.form.get("player_name")
-        player.player_country = request.form.get("player_country")
-        player.player_position = request.form.get("player_position")
-        player.player_joined = request.form.get("player_joined")
-        db.session.commit()
-        players = list(Player.query.order_by(Player.player_kit_number).all())
-        return render_template("players.html", players=players, team=team)
+        if session["user"] == team.team_created_by:
+            player.player_kit_number=request.form.get("player_kit_number")
+            player.player_name = request.form.get("player_name")
+            player.player_country = request.form.get("player_country")
+            player.player_position = request.form.get("player_position")
+            player.player_joined = request.form.get("player_joined")
+            db.session.commit()
+            players = list(Player.query.order_by(Player.player_kit_number).all())
+            return render_template("players.html", players=players, team=team)
+        else:
+            flash("Only the creator of this team may edit its players.")
     return render_template("edit_player.html", player=player, team=team, teams=teams)
 
 
@@ -178,20 +196,21 @@ def add_player(id):
             return render_template("add_player.html", teams=teams, team=team,)
     
     if request.method == "POST":
-        player = Player(
-            player_kit_number=int(request.form.get("player_kit_number")),
-            player_name=request.form.get("player_name"),
-            player_position=request.form.get("player_position"),
-            player_joined=request.form.get("player_joined"),
-            player_country=request.form.get("player_country"),
-            team_id=request.form.get("team_id")
-        )
-
-        db.session.add(player)
-        db.session.commit()
-        
-        players = list(Player.query.order_by(Player.player_kit_number).all())
-        return render_template("players.html", players=players, team=team)
+        if session["user"] == team.team_created_by:
+            player = Player(
+                player_kit_number=int(request.form.get("player_kit_number")),
+                player_name=request.form.get("player_name"),
+                player_position=request.form.get("player_position"),
+                player_joined=request.form.get("player_joined"),
+                player_country=request.form.get("player_country"),
+                team_id=request.form.get("team_id")
+            )
+            db.session.add(player)
+            db.session.commit()
+            players = list(Player.query.order_by(Player.player_kit_number).all())
+            return render_template("players.html", players=players, team=team)
+        else:
+            flash("Only the creator of this team may add players to it.")
 
     return render_template("add_player.html", teams=teams, team=team)
 
@@ -214,10 +233,13 @@ def team_profile(id):
 def delete_player(team_id, player_id):
     player = Player.query.get_or_404(player_id)
     team = Team.query.get_or_404(team_id)
-    db.session.delete(player)
-    db.session.commit()
-    players = list(Player.query.order_by(Player.player_kit_number).all())
-    return render_template("players.html", players=players, team=team)
+    if session["user"] == team.team_created_by:
+        db.session.delete(player)
+        db.session.commit()
+        players = list(Player.query.order_by(Player.player_kit_number).all())
+        return render_template("players.html", players=players, team=team)
+    else:
+        flash("Only the creator of this team may delete its players.")
 
 
 @app.route("/user_edit/<username>", methods=["GET", "POST"])
